pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "batman"
authors = [
    {name = "Nuclear RAMP team", email = "nuclear.ramp.dev@gmail.com"},
]
description = "Package for solving reactions and decays by Bateman's equation"
dynamic = ["version"]
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "dask",
    "matplotlib",
    "more-itertools",
    "multipledispatch",
    "networkx",
    "numpy",
    "ramp-coremaker",
    "ramp-endf",
    "ramp-isotopes",
    "ramp-reactions",
    "scipy",
    "toolz"
]

[project.optional-dependencies]
test = [
    "hypothesis",
    "pytest",
]

[tool.setuptools.packages.find]
include = ['batman*']
exclude = ['tests*']

[tool.setuptools_scm]
